% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macleish.R
\name{read_whately}
\alias{read_whately}
\alias{get_daily}
\title{Read in the current Whately and Orchard data}
\usage{
read_whately()

get_daily(x)
}
\description{
Read in the current Whately and Orchard data
}
\examples{
\dontrun{
mac_data <- read_whately()
names(mac_data)
lapply(mac_data, dim)
lapply(mac_data, head)
}
}
\seealso{
\code{\link[macleish]{whately_2015}}
}
