% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macleish.R
\name{read_whately}
\alias{read_whately}
\title{Read in the current Whately data}
\usage{
read_whately()
}
\description{
Read in the current Whately data
}
\examples{
\dontrun{
whately <- read_whately()
nrow(whately)
head(whately)
}
}
